feat: refactor MentorProgram Actions structure and refactor PrimaryButton component
